requirements.txt
Flask==2.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.5.0
Flask-JWT-Extended==4.4.4
Flask-Mail==0.9.1
Flask-Limiter==2.8.1
Flask-Caching==2.0.1
python-dotenv==0.19.2
bcrypt==4.0.1
redis==4.3.4
SQLAlchemy==2.0.38
phonenumbers==8.13.22
requests==2.31.0
Flask-WTF==1.1.1